# Conflicts:
#	GroundStation.qrc
#	src/AutoPilotPlugins/APM/APMSafetyComponentCopter.qml
#	src/AutoPilotPlugins/PX4/AirframeComponentSummary.qml
#	src/AutoPilotPlugins/PX4/FlightModesComponentSummary.qml
#	src/AutoPilotPlugins/PX4/PX4RadioComponentSummary.qml
#	src/AutoPilotPlugins/PX4/PX4TuningComponentPlane.qml
#	src/AutoPilotPlugins/PX4/PX4TuningComponentVTOL.qml
#	src/AutoPilotPlugins/PX4/PowerComponent.qml
#	src/AutoPilotPlugins/PX4/PowerComponentSummary.qml
#	src/AutoPilotPlugins/PX4/SafetyComponentSummary.qml
#	src/AutoPilotPlugins/PX4/SensorsComponentController.cc
#	src/AutoPilotPlugins/PX4/SensorsComponentSummary.qml
#	src/FirmwarePlugin/PX4/PX4FirmwarePlugin.h
#	src/FlightDisplay/FlightDisplayView.qml
#	src/FlightDisplay/FlightDisplayViewWidgets.qml
#	src/MissionManager/MissionController.cc
#	src/MissionManager/MissionController.h
#	src/MissionManager/MissionSettingsItem.h
#	src/MissionManager/SimpleMissionItem.cc
#	src/MissionManager/StructureScanComplexItem.cc
#	src/MissionManager/StructureScanComplexItem.h
#	src/MissionManager/Survey.SettingsGroup.json
#	src/MissionManager/SurveyMissionItem.cc
#	src/MissionManager/SurveyMissionItem.h
#	src/MissionManager/TransectStyleComplexItem.h
#	src/PlanView/CameraSection.qml
#	src/PlanView/CorridorScanEditor.qml
#	src/PlanView/MissionSettingsEditor.qml
#	src/PlanView/PlanView.qml
#	src/PlanView/SimpleItemEditor.qml
#	src/PlanView/SurveyItemEditor.qml
#	src/PlanView/SurveyMapVisual.qml
#	src/QmlControls/ParameterEditorDialog.qml
#	src/QmlControls/QGCTextField.qml
#	src/Vehicle/Vehicle.cc
#	src/Vehicle/Vehicle.h
#	src/VehicleSetup/FirmwareUpgrade.qml
#	src/comm/LinkManager.cc
#	src/comm/USBBoardInfo.json
#	src/uas/UAS.h
#	src/ui/QGCMAVLinkLogPlayer.cc

+++ src/QGC.h
@@ -36,7 +36,7 @@ quint64 groundTimeMilliseconds();
  */
 qreal groundTimeSeconds();
 /** @brief Returns the angle limited to -pi - pi */
-float limitAngleToPMPIf(float angle);
+float limitAngleToPMPIf(double angle);
 /** @brief Returns the angle limited to -pi - pi */
 double limitAngleToPMPId(double angle);
